Master Audio Slider
The Master Volume Slider detects the Master Audio Mixer but it doesn't seem to lower the background volume after I slide it over. This may be an issue with what the audios' output is in the inspector.
Added sound effects to the Cheat toggles. May also need to add audio mixer output to the button sounds.